+++ hutool-core/src/test/java/cn/hutool/core/date/ChineseDateTest.java
@@ -1,6 +1,5 @@
 package cn.hutool.core.date;
 
-import cn.hutool.core.lang.Console;
 import cn.hutool.core.util.StrUtil;
 import org.junit.Assert;
 import org.junit.Test;
@@ -118,10 +117,26 @@ public class ChineseDateTest {
 	}
 
 	@Test
-	public void dayTest(){
-		Date date = DateUtil.parse("1900-01-31");
-		//Date date = DateUtil.parse("2022-02-22","yyyy-MM-dd");
+	public void day19700101Test(){
+		// https://gitee.com/dromara/hutool/issues/I4UTPK
+		Date date = DateUtil.parse("1970-01-01");
 		ChineseDate chineseDate = new ChineseDate(date);
-		Console.log(chineseDate);
+		Assert.assertEquals("己酉鸡年 冬月廿四", chineseDate.toString());
+
+		date = DateUtil.parse("1970-01-02");
+		chineseDate = new ChineseDate(date);
+		Assert.assertEquals("己酉鸡年 冬月廿五", chineseDate.toString());
+
+		date = DateUtil.parse("1970-01-03");
+		chineseDate = new ChineseDate(date);
+		Assert.assertEquals("己酉鸡年 冬月廿六", chineseDate.toString());
+	}
+
+	@Test
+	public void day19000101Test(){
+		// 1900-01-31之前不支持
+		Date date = DateUtil.parse("1900-01-31");
+		ChineseDate chineseDate = new ChineseDate(date);
+		Assert.assertEquals("庚子鼠年 正月初一", chineseDate.toString());
 	}
 }